+++ libnodegl/node_compute.c
@@ -52,7 +52,6 @@ struct compute_priv {
                                           NGL_NODE_UNIFORMVEC3,     \
                                           NGL_NODE_UNIFORMVEC4,     \
                                           NGL_NODE_UNIFORMCOLOR,    \
-                                          NGL_NODE_UNIFORMCOLORA,   \
                                           NGL_NODE_UNIFORMQUAT,     \
                                           NGL_NODE_UNIFORMBOOL,     \
                                           NGL_NODE_UNIFORMINT,      \