+++ libnodegl/node_render.c
@@ -84,7 +84,6 @@ struct render_priv {
                                           NGL_NODE_UNIFORMVEC3,     \
                                           NGL_NODE_UNIFORMVEC4,     \
                                           NGL_NODE_UNIFORMCOLOR,    \
-                                          NGL_NODE_UNIFORMCOLORA,   \
                                           NGL_NODE_UNIFORMQUAT,     \
                                           NGL_NODE_UNIFORMINT,      \
                                           NGL_NODE_UNIFORMIVEC2,    \