+++ libnodegl/nodes_register.h
@@ -185,7 +185,6 @@
     action(NGL_NODE_UNIFORMVEC3,            ngli_uniformvec3_class)             \
     action(NGL_NODE_UNIFORMVEC4,            ngli_uniformvec4_class)             \
     action(NGL_NODE_UNIFORMCOLOR,           ngli_uniformcolor_class)            \
-    action(NGL_NODE_UNIFORMCOLORA,          ngli_uniformcolora_class)           \
     action(NGL_NODE_UNIFORMQUAT,            ngli_uniformquat_class)             \
     action(NGL_NODE_USERSELECT,             ngli_userselect_class)              \
     action(NGL_NODE_USERSWITCH,             ngli_userswitch_class)              \